Add Prettier
The config is copied from `consumer`, except for the stuff about
multiline arrays. Without that plugin/config, there were changes in
the JSON data files that I didn't like; I think it's best for arrays
in these data files to always be one item per line.

Most of this PR is prettier changes; the important bits are in
`package.json`, `.prettierrc`, `.prettierignore`.

Testing:

- Look at all the formatting changes, make sure they're not semantic

- Put a double-quoted string in a JS file, make sure `yarn test` fails