Factor out the list of items we have incentives for (#89)
This now allows verifying that the `item` key in the incentives data
is a valid item that we have locale strings for. `item_url` is removed
from the static data, as it's redundant. However, it still appears in
API responses (containing an actual URL).

The i18n code is now less general than before, but this allows for
robust typing.  If it turns out we need `t()` to have more general
capabilities, we'll cross that bridge when we come to it.