TypeScript conversion for v1 routes (#90)
This is a big one. The major change is adding Fastify's support for
typing requests based on schemas. This required moving the v1 JSON
Schemas into JS rather than JSON.

This breaks the manual workflow of copying the schema files into
Zuplo, but that's not a problem for now: the whole v1 API is not
schema-fied in Zuplo right now. If/when we need to do that, it should
be simple to write a script in this repo that outputs these schema
objects as JSON.