Fixture test for the v1 calculator API
This was commented out with a TODO because the API format is going to
change. That's still the case, but I'd rather have this test running,
because I'm making a bunch of code changes that I don't intend to
cause API changes, and it's a good guardrail.

I generated the fixture from the state of this repo just before I
started working on it, and compared it with the `v0` fixture. There
were only the known changes:

- `estimated_annual_savings` removed from the top level
- `_es` localized keys aren't included
- The relative `more_info_url` became the absolute `item_url`
- The solar tax credit is sorted in a different position

Then I ran the test to compare it with the current `v1`, and found a
change that I made unintentionally in #90, but that I think is
correct: when `agi_max_limit` and `representative_amount` are null in
the underlying data, they now come out as `null` in the API, rather
than `0` as before. This was because the corresponding entries in the
JSON schema for incentives in the API erroneously didn't include
`null` as a possible type for those fields.