Add /calculator param to filter by item
PEP will want this feature (as I've alluded to in Slack) and the
designs for the RI calculator assume it as well.

This assumes that the existing taxonomy of products/projects (which
comes from the IRA data) is what we want for this feature, and I'm not
sure that's the case long-term. This will be good enough for the time
being.

New unit test passes; a few spot-check manual requests return the
expected results.