Turn Errors from incentive calculation into HTTP 400
Uncaught generic exceptions turn into HTTP 500 responses, which isn't
what we want for these invalid-input cases.

I'm actually not sure this is the right approach. The other option
would be to throw `fastify.httpErrors.badRequest` directly from the
origin of the errors, in `incentive-calculation.ts`. That felt weird,
being presentation-layer logic in a business-logic module. But the
approach in this PR risks turning errors into bad-request responses
when they shouldn't be. Opinions welcome.