Switch items to snake case (#106)
As discussed in #105, now that these item names are exposed in the API
as part of the `items` filter, they should match the naming convention
of everything else in the API, which is snake case.

This is just straight find-and-replace. And if you forget to change
any one of these places, a test fails, which is nice.