Turn Errors from incentive calculation into HTTP 400 (#107)
Uncaught generic exceptions turn into HTTP 500 responses, which isn't
what we want for these invalid-input cases.

This throws different classes of error in different circumstances,
allowing the top-level code in `routes` to distinguish which HTTP
status it should return. It returns 400 for "invalid input" errors,
and rethrows anything else, letting Fastify turn it into a 500.

This actually caught a problem in the existing validation, which I've
now fixed.