Convert v0.js to TypeScript; turn off allowJs
## Description

This completes the TS conversion.

- The new files in `src/schema/v0` are copy-pastes of the now-deleted
  files in `schemas/v0`. The only substantive change is the addition
  of `additionalProperties: false`, which prevents `json-schema-to-ts`
  from adding a catchall `[x: string]: unknown` field to the object
  types.

- Adjustment to the `calculateIncentives` signature to make it
  smoother to call it from both v0 and v1 code.

## Test Plan

- `yarn test`
- Start server and hit each v0 endpoint

## Next Steps

I'd kinda like to switch to CommonJS modules. That's more the path of
least resistance for pure-Node projects. Now that everything is TS,
we'll still be able to use `import`/`export` syntax, but we won't need
the `.js` extensions in imports, and we can remove some of the hacks I
had to add to get Fastify and ts-node to work with ES modules.