Switch to CommonJS modules
## Description

This is the generally-expected config for pure Node projects. Node's
ESM support is not mature yet, and I had to do some slightly janky
stuff to get it to work.

Now that everything is TypeScript, though, it's easy to switch module
systems. The substantive changes are:

- `package.json`: adjusting the commands to load ts-node more cleanly,
  removing `"type": "module"`.

- `tsconfig.json`: changing to `commonjs`.

- `src/app.ts`: removing the `__dirname` shim and `forceESM` flag to
  fastify loading

- `test/helper.ts`: removing the `__dirname` shim

- Various tests: removing the `Ajv.default` workaround

Everything else is just removing suffixes from imports.

One interesting side effect of this is that `tap` checks code coverage
now, without me telling it to. I don't know exactly why this happened.
Anyway, by default it expects 100% test coverage (!!) and we don't
have that, so I've turned off the 100% enforcement for now. (It will
still print the coverage report at the end of each test run.)

## Test Plan

- `yarn test`

- `yarn dev`, run some manual requests with curl

## Next Steps

See what it would take to get 100% test coverage, I guess? If it
doesn't seem worthwhile, maybe figure out a reasonably attainable
coverage threshold and configure tap to enforce it?