Log the Authorization header on /api/v1/incentives (#116)
I'm trying to debug why the Zuplo service account isn't successfully
authenticating to the GCR backend, on prod only. (It's working in
dev!)

This authentication is supposed to happen by means of the
Authorization header, which contains a JWT. This should let us see
how/whether something's going wrong in that step of auth.

## Test Plan

- `yarn dev`, run curl locally without an Authorization header, and
  with a populated one; make sure logs are visible in both cases.

## Next Steps

Remove this after debugging.