Use enums everywhere possible
## Description

In starting to add the "married filing separately" status, I realized
that these lists of strings for the various enums were duplicated all
over the place. This turns all of them into enums.

There's a little wrinkle here that comes from two separate
schema-consuming libraries. It has to do with nullable enums:

```typescript
{
  type: 'string',
  nullable: true,
  enum: [...Object.values(FilingStatus), null]
}
```

It might seem redundant to have `nullable: true` and `null` in the
array, but:

- `ajv`, which does schema validation in unit tests, considers the
  `enum` array to be authoritative. If it doesn't include `null`
  explicitly, then `null` isn't a valid value there, regardless of the
  `nullable` flag.

- `json-schema-to-ts`, which generates TS types from schema, will
  _not_ allow `null` as a possible value without the `nullable:
  true`. It generates a field of type `FilingStatus | null` for the
  above.

So we gotta have both.

## Test Plan

- `yarn test`

- `yarn dev`, send some test requests.