Add married-filing-separately tax status (#123)
## Description

It turns out this status actually gets different brackets from single
filers, which is not what I had thought before. This is definitely
good to add!

This adds the new brackets and some unit testing to make sure that the
tax bracket data is well-formed. This actually uncovered a couple of
bugs.

The only existing incentive that cares about filing status are the
new/used EV tax credits; those are updated to reflect that they treat
MFS filers the same as single filers.

This is potentially a breaking change for clients of
`/api/v0/incentives`, because it will return a `filing_status` value
that they haven't seen before. (`/calculator` clients are unaffected
because they won't see the value in API responses unless they pass it
in requests.) I don't know how big of a problem this is, or if we
should proactively inform anyone?

## Test Plan

New & existing unit tests pass.

A few manual tests of API requests to make sure the new value is accepted.