Make item into a struct (#127)
## Description

The RI frontend design requires knowing which item an incentive is
for, so we need to start sending a machine-friendly stable identifier
for items.

I'm turning it into a struct instead of just adding a field because it
seems conceptually cleaner and more scalable. It also allows the
possibility of a `/v1/items` endpoint in future.

Side note: the `APIIncentiveNonLocalized` thing is starting to be a
little awkward. It's an intermediate stage of incentive
representation, halfway between the static-data format and the API
format. It still works OK but I think eventually it might work better
to have incentive calculation return `v1`-format data.

## Test Plan

`yarn test` with updated fixtures.