Unify GCR service name across dev and prod
## Description

This will make it easier to define Terraform config for the service,
as it's one less thing that will vary across environments.

I'm also going to tag each new image with the branch name, as well as
the commit hash. This is mainly for developer experience; you can look
at the list of images and see where the `main` and `production`
branches are, instead of just commit hashes.

I've already created a Cloud Run service under the new name in the dev
GCP project, so this can land on main smoothly.

## Test Plan

I've tested the individual commands manually, making sure that:

- You can pass `-t` twice to `docker build`, and it applies both tags

- Passing `-a` to `docker push` makes it push all the tags that are
  applied to the same image ID