Add "account_credit" and "rebate" incentive types
## Description

I'd coded all the RI incentives as `pos_rebate`, when in fact a lot of
them are not point-of-sale rebates. This introduces two additional
types that correspond to:

- Account credits (an amount credited against your future utility bills); and

- After-the-fact rebates, which I've just called `rebate` because
  that's the common understanding of what a rebate is. Up-front
  rebates (like the IRA rebates) are the exceptional case.

  We've considered various other names for this concept, like
  explicitly saying "after-the-fact rebate" (weird, no one says this),
  "mail-in rebate" (a common phrase, but in a lot of cases you don't
  actually have to mail anything in), and "check rebate" (suggestive
  of how it works, but implicitly excludes electronic payments).

I've updated the incentives whose mechanisms I'm sure about; there may
be others that need updating, though.

This doesn't affect `v0` as none of the IRA incentives are getting
these new types.

## Test Plan

`yarn test`.