Add program URLs
## Description

This is to support sending people directly to program pages from
incentive cards in the calculator.

I debated between associating URLs with _programs_ or with _specific
incentives_. I think the per-program approach makes sense, and it
doesn't preclude eventually adding per-incentive URLs.

The URLs are localizable, but are optional for each program.

## Test Plan

`yarn test`, which includes an added test to make sure URLs are valid.