Make authority_types work as a filter
## Description

This is preparation for implementing a reasonable experience for the
RI calculator when you don't enter a RI zip code.

In the first-party context, we'll want to just fall back to IRA
incentives only, with no state map outline and utility selector. That
means the frontend wants to always receive incentives from all levels,
with graceful behavior if there's no state or utility coverage.

New behavior:

- If you don't pass `authority_types` at all, you always get federal
  incentives, plus state incentives if we have them.

  - If you pass a value in `utility`, you also get incentives for that
    utility **if** it's a valid utility ID in the state of
    `location`. If we don't have coverage in that state, or if we do
    but we don't have the utility ID, that's an error.

- If you pass `authority_types`, we filter incentives to only those
  authority types.

  - If you have `utility` in `authority_types`, and you don't pass a
    value for `utility`, that's an error. The idea is that you're
    explicitly asking for utility incentives; you have to tell us
    which one.

  - However, if you have `state` in `authority_types`, it **is not**
    an error if we don't have coverage in that state. I could be
    convinced to go the other way on this.

## Test Plan

`yarn test` with updated tests.