Add logos to authorities
## Description

Logos are structs with width and height values for the images, rather
than just image URLs, so that frontends can render sized images and
thus avoid redraws after the images load.

One thing I considered is setting the width and height as metadata in
Cloud Storage, and having the API backend read that metadata to serve
to the frontend, rather than hardcoded the dimensions next to the
URL. But that seemed like a needless number of moving pieces.

## Test Plan

`yarn test` with updated fixtures