Get rid of APIIncentiveNonLocalized
## Description

I'd mentioned before that this concept was getting awkward, and we're
starting to hit that point. This is the fix.

The core change here is to have `calculateIncentives` return a type
that is a union of the IRA static-data type and the RI static-dat
type, plus the `eligible` flag.

- Renamed `Incentive` to `IRAIncentive` for clarity.

- Adding the `authority_type: federal` flag to the IRA data directly
  just makes things smoother. That means the union type `(Incentive |
  RIIncentive)` has the `authority_type` field required, which matches
  with the fact that it's required in `APIIncentive`, which just makes
  things smoother.

- `start_date` and `end_date` in the state incentives definition:
  they've been there in the JSON the whole time, just not expressed in
  the types.

## Test Plan

`yarn test`