Script to update all the fixtures from real output
## Description

When making a change to the API output, I run `yarn test` to see the
diff, and then if the diff is right, I want to update the fixture file
to match. Currently, this entails either updating it manually, or
putting together just the right API request to get that output.

This script contains explicit curl commands that generate each fixture
file. I'm sure there's a prettier way to do this, but this is the
simplest, and it doubles as a quick way to get curl commands to run
locally.

I had to tell Fastify to stop watching the `test` directory, otherwise
it restarts the server in response to the fixture file changing, and
the next `curl` command fails because the server isn't up yet.

## Test Plan

Run the script; make sure there are no changes to the
fixtures. Introduce a deliberate change to API output and run the
script again; make sure that change is reflected in the fixture files.