Hoist authority info into a top-level struct (#148)
## Description

This is in preparation for adding logos to authorities. I wanted a
top-level struct for that, since a logo is going to be a struct and I
didn't want that duplicated in every incentive.

The structs under the top-level `authorities` key now just have the
authority name, but they'll soon also have logo info.

## Test Plan

`yarn test` with updated fixtures.