Add Drive+ incentives to RI dataset. (#153)
Add Drive+ incentives for RI.

I modeled this as a separate, low-income only incentive that can stack
on top of the existing Drive EV incentives. The logic was simple and
this was how it read to me from the spreadsheet and the RI Drive website
(however, the short descriptions in the spreadsheet were inconsistent
with this model so those will have to be updated if we go this route).

I can also see an argument for modeling this as a **replacement** for
the Drive EV ones, but that felt a little more complicated since we'll
have to handle something like the case where we set low_income
explicitly to false and check for that in the incentives calculation.
I'm open to switching it to that.


Tested using `yarn test`. The total savings increase as expected for the
low income user.

(I should have branched this from where I was working on the low income
thresholds but I merged them and tested it out locally and this change
still works with that one.)