Add location to response
## Description

Unlike the `coverage` field I recently added, this is simply the
output of computing the location from the request's `location` field;
it's unrelated to which sets of incentives we evaluated for
eligibility.

This is what the frontend will actually need. When locked to a
specific state, it wants to show an error message whenever the entered
ZIP is not in the right state.

## Test Plan

yarn test