Don't error in /utilities if we don't have coverage (#167)
The frontend will rely on the `location` in this response to implement
locking to a specific state, so it needs to know the location
regardless of what utility coverage exists.

The semantics of this endpoint are now straightforwardly "here are the
utilities that we know may be operating in the given location". This
includes the possibility of "none", when we just don't know what
utilities are there.

NB: I think you could make a good argument for either behavior here.
This way happens to make the frontend logic a little cleaner, but I
don't know if that should be what drives the decision.