Add flags for beta and launched states (#166)
I decided to hard code a list of "launched" and "beta" states, rather
than including them as query parameters. This way you don't have to know
what states we're working on in order to ask for the ones that are still
in development. I can see the argument for doing it the other way,
though, so let me know if that would be preferable.


Tested with `yarn test`, it appears to be a no-op as expected.

I also tested by running it locally and issuing queries for different
zips and it behaved as expected, only returning state incentives for RI
zips.