Update Clean Heat RI JSON and unit tests (#186)
3 things I want to resolve before merging:

1) Do the new heat pump/electrical panel incentives apply to both
renters and homeowners, or homeowners only?
2) For the ones that are set with low_income = false, does that mean
they are NOT offered to households below the low income thresholds, or
does it mean not ONLY low income households? If so, we'll need a quick
update to be able to correctly support that (so, the geothermal one is
listed as low_income = false but we are returning it as an eligible
incentive for the low income test case).
3) RI-32 looks like it should technically be implemented as a bonus and
not a standalone incentive, but we don't support those yet. This is
probably ok to add for now but it should get an update once we add
support for bonuses and stacking (proposal
[here](https://docs.google.com/document/d/1qo6XbUODBN4kXE2Q9bqpJh1l1oIrIDkpmhADg2jGC5s/edit#heading=h.3hsarjfyiy96)).

Tested by updating the unit tests and checking the diffs. They look
correct pending the above questions.

---------

Co-authored-by: Tom Carden <tomc@rewiringamerica.org>