Remove 'glob' and update the required node.js version (#168)
I started by just removing the glob dependency, but in my local
environment `yarn` wouldn't run because the `.nvmrc` that I
automatically observe was set to v16 but the package.json requires v18.
Since the Dockerfile dictates what actually runs in Cloud Run, and that
was also v16, I was tempted to just change package.json to match. But
node v16 hit EOL last month, and node v18 is about to become
"Maintenance LTS" so it seems like we should be at least on v18.
https://nodejs.dev/en/about/releases/

So this PR has two commits:
 - set node versions to lts/hydrogen
 - remove direct glob dependency

I didn't investigate if nvm can set versions from package.json, nor did
I investigate if we can pull the version in the Dockerfile from one of
the other two files. Those would be nice enhancements if possible.

I think this is worth a little care to shepherd through some testing, so
it's in draft mode for the moment. Might be a good time for us to start
running the Docker build in CI?