Generalize state incentives schema (#191)
I'm hoping most of the states can have the same schema. We can always
create a specific one if there is a state that really doesn't fit in
with this.

Tested with unit tests and by running it locally and issuing some
queries for RI and making sure they still work.