Add first draft of CT incentives data (#175)
This is based on a [combination of the data
collection](https://docs.google.com/spreadsheets/d/1yS6Ypxyk0pL7T6arTKZaROqBF5EVbZc5yJRPne1chNY/edit#gid=557942691)
that Dan and I did.

I left out the ones that had complicated stacking or the two-part
incentives for now, since we can't correctly handle those until we have
the ability to define IDs for the incentives and relationships between
them.

Tested:
Added a unit test to check that CT can be included when the beta states
are requested. Also added a test for a low income CT user.