Schema for localizing program names and URLs
## Description

The `locales` files are starting to get unwieldy, so this splits
programs (names and URLs) into their own data file, and introduces a
`LocalizableString` schema that we can use for other localizable
strings.

The idea is to replace simple string properties with objects that look
like:

```json
{
  "en": "English",
  "es": "Español"
}
```

The `en` key is the only required one. If something calls for a
non-`en` version but there isn't one available, we fall back to `en`.

My intent is to do this for `short_description` too, if everybody is
on board with this approach.

## Test Plan

`yarn lint` and `yarn test`