Fix Tap-related type error
I don't know when this started happening (I haven't worked in this
codebase for a few weeks), but I was getting type errors running `tsc`
locally.

It turns out we had conflicting versions of `tap` and `@types/tap`
installed. The newest version of `tap` ships types, so the `@types`
version is no longer needed.