Localizable short descriptions (#208)
## Description

Update all the `short_description`s in the data to use the
localizable-string schema I introduced in #207.

`short_description` is now required to be present for all incentives.
Since it's no longer possible to validate the length of the strings
using JSON schema, I added a unit test to do that.

Spanish translations exist in a spreadsheet for the IRA incentives;
I'll add those in a followup, to keep the data changes in this PR
mechanical.

## Test Plan

`yarn test`. Deliberately make a description too long and make sure
the test fails.