Add Spanish IRA descriptions + script to update them (#210)
## Description

I updated `update-ri-descriptions.ts` to add support for files other
than the RI incentives, with the aim of using it to populate the
Spanish descriptions for IRA incentives.

One thing that's a little weird here is that there are multiple
entries in the IRA JSON that have the same ID. That's because they're
about the same underlying incentive (and thus the same row in the
spreadsheet), but they're expressing different amounts for different
income levels or tax statuses or whatever. I'm not sure how this has
been handled in the state incentive sheets, so I'm happy to change
as needed.

## Test Plan

Run the script with `RI` as the argument; make sure there are no
changes. Ran it with `IRA` as the argument to make the included
changes to the IRA data.