Github Actions workflow to deploy
## Description

The summary is:

- On **pull request** creation or update, run tests
- On **push to main**, run tests and deploy to _staging_
- On **release**, run tests and deploy to _prod_

It all happens using off-the-shelf Github Actions. The necessary setup was to create a service account for the workflow to act as (rewiringamerica/terraform#34) and put the service account JSON keys in the Github secrets config for this repo.

I'm taking this opportunity to switch from the Container Registry (deprecated) to Artifact Registry.

With this, we'll completely stop using Cloud Build for this service. A Github runner will build the image and push it to Artifact Registry.

## Test Plan

Tested the workflow locally, as much as possible, using https://github.com/nektos/act.