Rewrite README for open source
## Description

This is a from-scratch rewrite of the README, with the target audience of an external developer or potential consumer of the API. There's also a CONTRIBUTING.md with a guide to the codebase.

Information specific to RA's deployment of the code has been moved to our Notion instance, linked from the "Incentives / Savings Calculator" page.

### Next steps

- Make issue templates for reporting the situations identified in the "contributing" section

- (Maybe) write a primer on incentives and all the various jargon in here?

- License

## Test Plan

read it