Basic set up for incentive calculation unit tests (#221)
For now I just added a very basic test to set up the infrastructure for
this. After we've settled on the implementation, I'll use it to add some
more test data for validating my incentive prerequisite code.

@oyamauchi Especially let me know if you have concerns about me adding
the parameter to calculateStateIncentivesAndSavings(). I don't think
it's risky/brittle but I could be wrong.

Tested with `yarn test`.