Add license
## Description

One question is whether we want to add a license header to source
files. It would look like:

```ts
/*
 * Copyright 2023 Windward Fund
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
```

or a shorter version:

```ts
// Copyright 2023 Windward Fund
// SPDX-License-Identifier: Apache-2.0
```

According to Apache Foundation, this is optional but recommended. The
license applies without it, but it provides clarity in case source
files somehow start floating around separate from the rest of the
repo. I'm inclined not to bother, but I'm not the best decider here.

## Test Plan

Look at it