Remove $3k cap on Clean Heat RI low-income HP incentive
## Description

The incentive is uncapped. The $3k is additional money available for an electrical service upgrade. Ideally we'd show this as a bonus incentive, but we don't have support for that yet, so we have to settle for saying it in the description.

## Test Plan

`yarn test`

Point frontend at my local API server, look up 02861 / $10,000, make sure the "100% of cost of a heat pump" incentive appears.