Add include_beta_states flag to utilities endpoint (#235)
## Description

Now that we have a fallback to returning all of a state's utilities
from `authorities.json` in the absence of state-specific logic in
`utilities-for-location.ts`, we need a way to gate states. Makes sense
to use the same logic as `/calculator`.

## Test Plan

`yarn test`. Locally request utilities for CT zip code 06033 with and
without the flag. Make sure results are the same for RI (02861) with
and without the flag.