CO JSON prework: add payment types (#237)
The larger change here is adding a payment_methods field to the API
which is the successor of the type field. It's the same data but
there can be more than one value and with a more informative name. We
are not removing the current field in this PR, just adding the new one.
We'll update the frontend and then delete the old field from the API
later.

I have *not* changed the v0 API since my impression is that we want that
API to be frozen, but the underlying IRA data is updated here.