Initial implementation of incentive prerequisite relationships (#223)
This is the implementation of prerequisite relationships for incentives,
as proposed
[here](https://docs.google.com/document/d/1qo6XbUODBN4kXE2Q9bqpJh1l1oIrIDkpmhADg2jGC5s/edit#heading=h.3hsarjfyiy96).

Once we agree that this is the way we want to do it, I'll go back and
add the exclusion relationships as well.

It's worth noting that this only affects the incentive eligibility and
savings calculation, so we'll have to do a bit more work to label the
bonus incentives if we want them to be displayed differently to the
user.

Testing:

- I added a cycle detection test to the schema tests, to make sure that
we don't accidentally add relationships JSON with circular dependencies
and cause something to crash.

- I also added some dummy incentives and relationships to test that this
works at a basic level. When I implement the exclusion relationships, I
can add a more complicated example.

- I also added a fake set of data that does include a cycle, just to
confirm that the cycle detection works.