Remove ev_charger_credit from item_type (#242)
## Description

There was only one incentive, in all the data, with this item_type:
the federal EV charger credit. It's also the only incentive in the IRA
data with `item == electric_vehicle_charger`, so it's easily
distinguished that way.

The v0 output is unchanged. The v1 output is changed, but the state
calculator never reads this field, so this is fine.

The final step of removing `item_type` will be dealing with
`performance_rebate`, which will require a slightly different treatment.

## Test Plan

`yarn test`