Add performance_rebate to PaymentMethod
## Description

This is the last step before removing `item_type`. This was the last
usage of the field in a way that couldn't be replicated with `type`
(soon to be `paymentMethods`).

## Test Plan

`yarn test`. Searching the code for `item_type` shows there are no remaining places that read the value.

The embed codebase (the only usage of v1) already includes
`performance_rebate` in the possible values of `type`, and handles it
properly: https://github.com/rewiringamerica/embed.rewiringamerica.org/blob/d764954497776851f9c3a2cce1e6e663669e5527/src/state-incentive-details.ts#L343