Add mutual exclusion relationships logic (#240)
Things get more complicated when we have two types of relationships,
since it becomes possible for incentives to switch back and forth
between eligible and ineligible multiple times before everything is
settled. I tried to come up with reasonably complicated test cases to
ensure that this works but I'd be happy to add more if there is a type
of example I missed.

I don't love the naming of requires/requiredBy/supersedes/supersededBy
but didn't want to block on this for too long. I'd be happy to hear
suggestions for better names.