Use parent ZCTA when available
## Description

The distinction between ZIPs that are ZCTAs and ZIPs that aren't keeps
confusing internal testers, and it's a decent bet that external users
have gotten bit by this too.

The `location` field in the v1 response doesn't currently contain a
ZIP, but there's an opportunity here to return the ZIP we used for
calculations in case it differs from the one the user entered. I don't
feel very strongly about this, and I think maybe it could stand to
wait until we add a more thought-out "showing our work" feature.

https://app.asana.com/0/1204738794846444/1205397060691934

## Test Plan

New and updated tests.