Add logic for combined maximum value across groups of incentives (#251)
My current thinking is that an incentive should only be allowed to be in
a single grouping, since otherwise the order of evaluation could
influence the resulting total savings number. This would be a confusing
user experience since we don't even know what order they'd be likely to
claim incentives in. It will require the code to be a lot more complex
as well. I think this is an ok data constraint to enforce since most of
the time all of the incentives in a group will likely be coming from the
same program/authority so a single incentive shouldn't need to be in
multiple groups anyway.

**Test plan**
I added some unit tests for different grouping scenarios. I also added
the CT incentives that were part of groups and confirmed some of them
show up for the CT test user, although they don't currently contribute
to the total savings values since they don't have representative values.