Generate valid OpenAPI (#247)
## Description

The main problem was that almost nowhere we had an `"example"` field
is actually allowed to have it. OpenAPI 3 only allows them in a few
places:

- Parameters

- Headers

- "Media Type": that is, here:
  ```
  paths:
    /api/v0/calculator:
      get:
        responses:
	  200:
	    content:
	      application/json:
	        schema: ...
		examples: # here
  ```
However, I haven't found a way to get `fastify-swagger` to put examples
there.


https://raw.githubusercontent.com/OAI/OpenAPI-Specification/main/schemas/v3.0/schema.json

For the most part I think getting rid of the examples is fine. Some of
them are trivial -- you don't need an example of a number between 1
and 8. None of the parameters are hard to understand.

The examples of entire incentives had gotten out of date because they
aren't checked. (JSON Schema doesn't actually require that `examples`
conform to the schema they're in; it's only "recommended".) I think
there's nonzero value in those, but we can define them in the Zuplo
config (maybe as part of some automation to keep the schemas in Zuplo
up to date?)

For the v0 incentive fields, instead of examples I wrote
`description`s. I didn't do that for v1 because I anticipate doing a
thorough pass over v1 when we start to open it up externally.

One other error was that you can't have any properties alongside a
`$ref`.

## Test Plan

`curl http://localhost:3000/spec.json` and paste it into:

- https://validator.swagger.io/
- https://apitools.dev/swagger-parser/online/

and get no errors.

`yarn test`


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1206040063611968