Updating CT incentives to use relationships (#253)
As part of this, I ended up reviewing some of the source data and found
that some of the incentives had changed, so I updated those values
accordingly and deleted the incentives that seem to not exist anymore. I
updated the spreadsheet to match. Unfortunately it's kinda
ugly/confusing to review in github :(

**Test plan**
Updated the CT test case. The user is eligible for the set of Eversource
incentives, as well as the new bonus ones that I just added.