Add concept of permanently ineligible incentives (#266)
If an incentive was marked as ineligible before relationships were
checked (e.g. due to income, owner status, etc), the relationships logic
should not be allowed to mark it as eligible.

**Test plan**
Added a unit test for a case where an income-ineligible incentive would
have been accidentally marked as eligible by the relationships logic
previously.