Split local incentives into city and county types and add support for serving them. (#272)
City and County are different kinds of local authorities and the logic
for matching them is different, so we split them here and add support
for serving them in the API (previously they've been filtered out).

Summary of algorithm:
1. When looking up location information, use the zip code to return a
county and city for the request, not just a state ID.
2. Require that for any incentive with a county/city authority type,
that authority has county and possibly city attributes defined in
authorities.json. This is enforced in tests.
3. For county incentives, the request county and authority county need
to match. For city incentives, both the county *and* city attributes
need to match, because there can be multiple cities with the same name
in a state.

I can add documentation in a README about how to do #2, but want to make
sure we are okay with this approach first.